documentation for the rest of subclasses of EventStream
